refactor: move onChange test to separate file
onChange test requires import from preact/compat which affects other tests in the same test file so I moved it to separate file
